+++ Lunar/Post Views/PostsView.swift
@@ -158,6 +158,7 @@ struct PostsView: View {
         Markdown { communityDescription ?? "" }
           .markdownTextStyle(\.text) { FontSize(fontSize) }
           .markdownTheme(.gitHub)
+          .markdownImageProvider(.lazyImageProvider)
       } label: {
         HStack {
           communityIconView
@@ -202,6 +203,7 @@ struct PostsView: View {
         Markdown { personBio ?? "" }
           .markdownTextStyle(\.text) { FontSize(fontSize) }
           .markdownTheme(.gitHub)
+          .markdownImageProvider(.lazyImageProvider)
       } label: {
         HStack {
           personAvatarView